requirements.txt
asgiref==3.7.2
bagit==1.8.1
Beaker==1.12.1
boltons==23.1.1
bullet==2.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.4.4
distlib==0.3.7
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
edam-ontology==1.25.2
futures==3.0.5
GDAL==3.5.2
idna==3.6
illusional.wdlgen==0.3.1
inflection==0.5.1
lockfile==0.12.2
markdown2==2.4.13
mistune==0.8.4
mock-server==0.3.7
nose==1.3.7
numpy==1.23.4
packaging==24.0
Parsley==1.3
Pillow==10.1.0
py-bcrypt==0.4
pygtail==0.14.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
repoze.lru==0.7
requests==2.31.0
ruamel.yaml==0.16.5
shellescape==3.8.1
sortedcontainers==2.4.0
sqlitedict==2.1.0
sqlparse==0.4.4
Tempita==0.5.2
tornado==4.2
typing_extensions==4.9.0
ubiquerg==0.6.3
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.9
webencodings==0.5.1
whitenoise==6.6.0
Whoosh==2.7.4
zipstream-new==1.1.8
